+++ client/src/app/shared/guards/redirect/redirect.guard.ts
@@ -1,0 +1,20 @@
+import {Injectable} from '@angular/core';
+import {ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot} from '@angular/router';
+import {StateService} from '../../services/state/state.service';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class RedirectGuard implements CanActivate {
+  constructor(
+    private state: StateService,
+  ) {}
+
+  canActivate(
+    route: ActivatedRouteSnapshot,
+    next: RouterStateSnapshot
+  ) {
+    this.state.entryPath = next.url;
+    return true;
+  }
+}